fix: 调整swebench_pro_infer.py中的build_instance函数
